% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cement.R
\name{cement}
\alias{cement}
\title{Function that concatenates strings without the objects needing to be quoted}
\usage{
cement(...)
}
\arguments{
\item{...}{objects to concatenate}
}
\value{
A concatenated string. Function pulled from https://adv-r.hadley.nz/quasiquotation.html
}
\description{
Function that concatenates strings without the objects needing to be quoted
}
