% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{Constants}
\alias{Constants}
\title{Ggplot chart constans}
\format{
A data frame with 9 rows and 9 variables
\describe{
\item{Names}{Name of constant}
\item{Values}{Value of constant}
}
}
\usage{
Constants
}
\description{
A data frame of constants use for my ggplots
}
\keyword{datasets}
