% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.choose}
\alias{model.choose}
\title{Calculator of AICc of nonlinear models from the drc package}
\usage{
model.choose(x, y, starters)
}
\arguments{
\item{x}{Independent cariable bector}

\item{y}{Dependent variable vector}

\item{starters}{Vector with names of starter functions}
}
\value{
Data frame of AICc values belonging to models from the DRC package
}
\description{
Calculator of AICc of nonlinear models from the drc package
}
