% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg.R
\name{summarised.tbl}
\alias{summarised.tbl}
\title{Function that automatically calculates the mean, standard deviation and standard error of desired variables}
\usage{
summarised.tbl(df.name, group, columns)
}
\arguments{
\item{df.name}{name of data frame to be rearranged as a string}

\item{group}{vector of strings with names of the columns you wish to group by}

\item{columns}{vector of strings with names of the columns you wish to summarise}
}
\value{
A data frame with specified means, standard deviations and standard errors
}
\description{
Function that automatically calculates the mean, standard deviation and standard error of desired variables
}
